Allow custom method names other than the default on server side
This new feature is usefull for a user that wants to use dot ('.') in
the xmlrpc method names available in its xmlrpc server.

Indeed, this is not possible in python function definition.
Therefore, this feature can be used to "simulate" nested methods.

The implementation of this feature does not use true nested methods
(i.e. in subclasses of the main class, inheriting from handler.XMLRPCView).

Instead, it simply allows the user to overwrite the method names thanks
to a `METHOD_NAMES` dictionary attribute in the user class (the one
inheriting from `handler.XMLRPCView`).

This approach has been used in order to not modify the whole code
structure of the 'handler.py' and still offer 'nested' methods from the
client point of view.

The README.rst and the tests have been updated and completed
accordingly.

Signed-off-by: Armand Bénéteau <armand.beneteau@iot.bzh>